+++ app/sections/AboutSection.js
@@ -1,0 +1,21 @@
+import { Container, Typography, Grid, Box } from '@mui/material';
+
+export default function AboutSection() {
+  return (
+    <header className="hero">
+      <div className="hero-content" id="about">
+        <p className="tagline">Hi, I am</p>
+        <h2>Abdul Kadir Khan</h2>
+        <p className="role">I am a Software Engineer</p>
+        <p>
+          I am an enthusiastic and adaptable individual, constantly seeking
+          out new opportunities for growth. Fueled by a strong desire to
+          expand my knowledge, I am committed to producing top-notch outcomes.
+          With an optimistic outlook and a determination to embrace
+          challenges, I am poised to make a valuable impact and attain
+          remarkable accomplishments.
+        </p>
+      </div>
+    </header>
+  );
+}